Frontend: Move recipient related code to api/projects module
